server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
numpy==1.24.3
pandas==2.1.4
pydantic>=2.6.1,<3.0.0
pydantic-settings>=2.1.0
python-dotenv==1.0.0
openai>=1.13.3,<2.0.0
python-multipart==0.0.9
crewai==0.28.8
crewai-tools==0.1.6
langchain>=0.1.10,<0.2.0
langchain-openai>=0.1.7,<0.2.0
google-search-results==2.4.2
newsapi-python==0.2.7
httpx==0.27.0